refactor: create labs feature directory (Phase 5c)
Move Labs components, store, and definitions to a new feature-based
directory structure:

- src/features/labs/components/
  - LabsButton, LabsDrawer, FeatureCard
  - FeatureStatusBadge, GraduatedSection, icons
- src/features/labs/store/labs.ts
- src/features/labs/definitions/{types,features}.ts
- Barrel exports at all levels

Re-export stubs maintain backward compatibility at old locations:
- src/components/Labs/
- src/core/labs/
- src/core/store/labs.ts

These re-export stubs will be removed in Phase 6.

Part of folder structure reorganization Phase 5.
