+++ src/features/labs/components/index.ts
@@ -1,0 +1,7 @@
+// Barrel export for labs components
+export { LabsButton } from './LabsButton';
+export { LabsDrawer } from './LabsDrawer';
+export { FeatureCard } from './FeatureCard';
+export { FeatureStatusBadge } from './FeatureStatusBadge';
+export { GraduatedSection } from './GraduatedSection';
+export { SparklesIcon, CloseIcon, ChevronRightIcon } from './icons';
