+++ src/features/labs/definitions/index.ts
@@ -1,0 +1,3 @@
+// Barrel export for labs definitions
+export * from './types';
+export * from './features';
