+++ src/features/labs/index.ts
@@ -1,0 +1,4 @@
+// Barrel export for labs feature
+export * from './components';
+export * from './store';
+export * from './definitions';
