+++ src/features/labs/store/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for labs store
+export { useLabsStore, LABS_STORAGE_KEY } from './labs';
